Update README.md
disclaimer: This is not a real life project as the data is very arbitary in nature.
created a formula or model to predict if a person is insured or not based on their age taking reference from past data.In this model we can see that most of the people who are aged over 40 are typically insured and the rest are not though there are some exception. Here  the input type is univariate type as  here if the person is insured or not is only depends on their age. In the plot area the values come either in 0 or in 1 (discrete), so, here we found a pattern or odel like here age>40 insured and age<40 not insured.